@@ -6,15 +6,15 @@ publish_to: 'none'
 version: 0.0.1
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^1.0.2
-  flame: ^1.0.0-rc5
-  flame_forge2d: ^0.6.0-rc2
+  cupertino_icons: ^1.0.3
+  flame: ^1.0.0-releasecandidate.14
+  flame_forge2d: ^0.8.1-releasecandidate.13
 
 dev_dependencies:
   flutter_test:
