@@ -1,7 +1,6 @@
 import 'dart:ui';
 
 import 'package:flame/sprite.dart';
-import 'package:flame/extensions/vector2.dart';
 import 'package:flame_forge2d/flame_forge2d.dart';
 import 'package:flame_forge2d/sprite_body_component.dart';
 import 'package:forge2d/forge2d.dart';
@@ -10,11 +9,11 @@ import 'package:game_testing/models/bullet_model.dart';
 /// The component of bullet
 class BulletComponent extends SpriteBodyComponent {
   final Bullet bullet;
-  Rect rect;
-  double startY;
+  late Rect rect;
+  late double startY;
   bool needRemove = false;
 
-  BulletComponent(Sprite sprite, this.startY, {Bullet bullet})
+  BulletComponent(Sprite sprite, this.startY, {Bullet? bullet})
       : this.bullet = bullet ?? Bullet.simple(),
         super(sprite, Vector2(11, 5));
 
@@ -27,10 +26,9 @@ class BulletComponent extends SpriteBodyComponent {
       Vector2(size.x, -size.y) / 2,
       Vector2(-size.x, -size.y) / 2,
     ];
-    shape.set(vertices, vertices.length);
-    final fixDef = FixtureDef()
+    shape.set(vertices);
+    final fixDef = FixtureDef(shape)
       ..userData = this
-      ..shape = shape
       ..restitution = 1.0
       ..density = 0.0
       ..friction = 0.0;
@@ -39,7 +37,7 @@ class BulletComponent extends SpriteBodyComponent {
     final def = BodyDef()
       ..userData = this
       ..position = viewport.getScreenToWorld(Vector2(40, startY))
-      ..type = BodyType.DYNAMIC;
+      ..type = BodyType.dynamic;
     this.startY = def.position.y;
     final s = viewport.size / viewport.scale;
     rect = Rect.fromLTWH(-s.x / 2, -s.y / 2, s.x, s.y);
