@@ -1,21 +1,22 @@
 import 'dart:ui';
 
-import 'package:flame/anchor.dart';
-import 'package:flame/components/text_component.dart';
-import 'package:flame/extensions/vector2.dart';
-import 'package:flame/text_config.dart';
+import 'package:flame/components.dart';
 
 /// Component that displays text health of tower.
 /// Health manipulated with [damageTower].
 class TowerHealthComponent extends TextComponent {
   final Vector2 startPosition;
   final double maxHealth;
-  Vector2 boxSize;
-  double currentHealth;
+  late Vector2 boxSize;
+  late double currentHealth;
 
   TowerHealthComponent(this.maxHealth, this.startPosition, double width)
-      : super(maxHealth.toStringAsFixed(1)) {
-    config = TextConfig(fontSize: 18, color: Color(0xFFFFFFFF));
+      : super(
+          maxHealth.toStringAsFixed(1),
+          textRenderer: TextPaint(
+            config: TextPaintConfig(fontSize: 18, color: Color(0xFFFFFFFF)),
+          ),
+        ) {
     position = startPosition + Vector2(width / 2, 10);
     anchor = Anchor.center;
     boxSize = Vector2(width, 20);
